Add `:legacy:search` module and move `SearchField`, `SearchAttribute` and `SearchCondition`
