+++ legacy/core/build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
     api(projects.legacy.di)
     api(projects.legacy.folder)
     api(projects.legacy.notification)
+    api(projects.legacy.search)
 
     implementation(projects.plugins.openpgpApiLib.openpgpApi)
