+++ legacy/core/src/main/java/com/fsck/k9/search/AccountSearchConditions.kt
@@ -2,10 +2,10 @@ package com.fsck.k9.search
 
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.Account.FolderMode
+import app.k9mail.legacy.search.api.SearchAttribute
+import app.k9mail.legacy.search.api.SearchCondition
+import app.k9mail.legacy.search.api.SearchField
 import com.fsck.k9.mail.FolderClass
-import com.fsck.k9.search.SearchSpecification.Attribute
-import com.fsck.k9.search.SearchSpecification.SearchCondition
-import com.fsck.k9.search.SearchSpecification.SearchField
 
 /**
  * Modify the supplied [LocalSearch] instance to limit the search to displayable folders.
@@ -17,16 +17,16 @@ fun LocalSearch.limitToDisplayableFolders(account: Account) {
     when (account.folderDisplayMode) {
         FolderMode.FIRST_CLASS -> {
             // Count messages in the INBOX and non-special first class folders
-            and(SearchField.DISPLAY_CLASS, FolderClass.FIRST_CLASS.name, Attribute.EQUALS)
+            and(SearchField.DISPLAY_CLASS, FolderClass.FIRST_CLASS.name, SearchAttribute.EQUALS)
         }
         FolderMode.FIRST_AND_SECOND_CLASS -> {
             // Count messages in the INBOX and non-special first and second class folders
-            and(SearchField.DISPLAY_CLASS, FolderClass.FIRST_CLASS.name, Attribute.EQUALS)
+            and(SearchField.DISPLAY_CLASS, FolderClass.FIRST_CLASS.name, SearchAttribute.EQUALS)
 
             // TODO: Create a proper interface for creating arbitrary condition trees
             val searchCondition = SearchCondition(
                 SearchField.DISPLAY_CLASS,
-                Attribute.EQUALS,
+                SearchAttribute.EQUALS,
                 FolderClass.SECOND_CLASS.name,
             )
             val root = conditions
@@ -38,7 +38,7 @@ fun LocalSearch.limitToDisplayableFolders(account: Account) {
         }
         FolderMode.NOT_SECOND_CLASS -> {
             // Count messages in the INBOX and non-special non-second-class folders
-            and(SearchField.DISPLAY_CLASS, FolderClass.SECOND_CLASS.name, Attribute.NOT_EQUALS)
+            and(SearchField.DISPLAY_CLASS, FolderClass.SECOND_CLASS.name, SearchAttribute.NOT_EQUALS)
         }
         FolderMode.ALL, FolderMode.NONE -> {
             // Count messages in the INBOX and non-special folders
@@ -66,12 +66,18 @@ fun LocalSearch.excludeSpecialFolders(account: Account) {
     this.excludeSpecialFolder(account.sentFolderId)
 
     account.inboxFolderId?.let { inboxFolderId ->
-        or(SearchCondition(SearchField.FOLDER, Attribute.EQUALS, inboxFolderId.toString()))
+        or(
+            SearchCondition(
+                SearchField.FOLDER,
+                SearchAttribute.EQUALS,
+                inboxFolderId.toString(),
+            ),
+        )
     }
 }
 
 private fun LocalSearch.excludeSpecialFolder(folderId: Long?) {
     if (folderId != null) {
-        and(SearchField.FOLDER, folderId.toString(), Attribute.NOT_EQUALS)
+        and(SearchField.FOLDER, folderId.toString(), SearchAttribute.NOT_EQUALS)
     }
 }
