+++ legacy/search/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id(ThunderbirdPlugins.Library.android)
+    alias(libs.plugins.kotlin.parcelize)
+}
+
+android {
+    namespace = "app.k9mail.legacy.search"
+}
