+++ legacy/search/src/main/java/app/k9mail/legacy/search/api/SearchAttribute.java
@@ -1,0 +1,19 @@
+package app.k9mail.legacy.search.api;
+
+
+///////////////////////////////////////////////////////////////
+// ATTRIBUTE enum
+///////////////////////////////////////////////////////////////
+public enum SearchAttribute {
+    CONTAINS,
+    NOT_CONTAINS,
+
+    EQUALS,
+    NOT_EQUALS,
+
+    STARTSWITH,
+    NOT_STARTSWITH,
+
+    ENDSWITH,
+    NOT_ENDSWITH
+}
